refactor(chisel): remove libs relative include paths
